\name{computePSNR}
\alias{computePSNR}

\title{
Utilities Bitmaps. Compute PSNR.
}
\description{
Computes the PSNR.
}
\usage{
computePSNR(x, y, norm = 255)
}

\arguments{
  \item{x}{
signal 1.
}
  \item{y}{
signal 2.
}
  \item{norm}{
the range, 1 or 255.
}
}
\value{
Returns the PSNR value.
}
\author{
Erwin Quiring.
}
